refactor: move realtime data collector
